requirements.txt
pygame-ce>=2.5.7
pillow>=10.0.0
